重写 AHCI 驱动, 更改 vdisk 接口

+++ src/x86_64/driver/pci/ide.c
@@ -446,12 +446,12 @@ void ide_write_sectors(uint8_t drive, uint8_t numsects, uint32_t lba, uint16_t e
     }
 }
 
-static size_t Read(int drive, uint8_t *buffer, uint32_t number, uint32_t lba) {
+static size_t Read(int drive, uint8_t *buffer, size_t number, size_t lba) {
     ide_read_sectors(drive_mapping[drive], number, lba, 1 * 8, (uint32_t)buffer);
     return number * SECTORS_ONCE;
 }
 
-static size_t Write(int drive, uint8_t *buffer, uint32_t number, uint32_t lba) {
+static size_t Write(int drive, uint8_t *buffer, size_t number, size_t lba) {
     ide_write_sectors(drive_mapping[drive], number, lba, 1 * 8, (uint32_t)buffer);
     return number * SECTORS_ONCE;
 }
