+++ src/x86_64/include/vdisk.h
@@ -10,14 +10,14 @@ typedef enum {
 } vdisk_flag_t;
 
 typedef struct {
-    size_t (*read)(int drive, uint8_t *buffer, uint32_t number, uint32_t lba);
-    size_t (*write)(int drive, uint8_t *buffer, uint32_t number, uint32_t lba);
+    size_t (*read)(int drive, uint8_t *buffer, size_t number, size_t lba);
+    size_t (*write)(int drive, uint8_t *buffer, size_t number, size_t lba);
     int (*ioctl)(size_t req, void *handle);
     int (*poll)(size_t events);
     void *(*map)(int drive, void *addr, uint64_t len);
     int          flag;
-    uint32_t     size;        // 大小
-    uint32_t     sector_size; // 扇区大小
+    size_t       size;        // 大小
+    size_t       sector_size; // 扇区大小
     vdisk_flag_t type;
     char         drive_name[50];
 } vdisk; // 块设备
@@ -32,6 +32,6 @@ int      vdisk_init();
  */
 int    regist_vdisk(vdisk vd);
 bool   have_vdisk(int drive);
-size_t vdisk_read(uint32_t lba, uint32_t number, void *buffer, int drive);
-size_t vdisk_write(uint32_t lba, uint32_t number, const void *buffer, int drive);
+size_t vdisk_read(size_t lba, size_t number, void *buffer, int drive);
+size_t vdisk_write(size_t lba, size_t number, const void *buffer, int drive);
 void  *device_mmap(int drive, void *addr, uint64_t len);
